+++ src/components/ids-breadcrumb/ids-breadcrumb.ts
@@ -159,7 +159,7 @@ export default class IdsBreadcrumb extends Base {
       <div class="ids-breadcrumb${truncated}">
         <div class="ids-breadcrumb-menu hidden">
           <ids-menu-button id="icon-button" menu="icon-menu">
-            <ids-icon slot="icon" icon="more"></ids-icon>
+            <ids-icon icon="more"></ids-icon>
             <span class="audible">Icon Only Button</span>
           </ids-menu-button>
           <ids-popup-menu id="icon-menu" target="#icon-button" trigger-type="click">