+++ src/components/ids-data-grid/ids-data-grid-formatters.ts
@@ -144,7 +144,7 @@ export default class IdsDataGridFormatters {
     // Type / disabled / icon / text
     return `<ids-button tabindex="-1" ${this.#columnDisabled(index, value, columnData, rowData) ? ' disabled="true"' : ''}${columnData.type ? ` type="${columnData.type}"` : ' type="tertiary"'}>
       <span class="audible">${columnData.text || ' Button'}</span>
-      ${columnData.icon ? `<ids-icon slot="icon" icon="${columnData.icon}"></ids-icon>` : ''}
+      ${columnData.icon ? `<ids-icon icon="${columnData.icon}"></ids-icon>` : ''}
     </ids-button>`;
   }
 
@@ -161,7 +161,7 @@ export default class IdsDataGridFormatters {
   tree(rowData: Record<string, unknown>, columnData: IdsDataGridColumn): string {
     const value: any = this.#extractValue(rowData, columnData.field);
     const button = rowData?.children ? `<ids-button tabindex="-1" class="expand-button">
-      <ids-icon slot="icon" icon="plusminus-folder-${rowData.rowExpanded === false ? 'closed' : 'open'}"></ids-icon>
+      <ids-icon icon="plusminus-folder-${rowData.rowExpanded === false ? 'closed' : 'open'}"></ids-icon>
     </ids-button>` : '&nbsp;';
     return `<span class="ids-data-grid-tree-container">${button}<span class="text-ellipsis">${value}</span></span>`;
   }
@@ -170,7 +170,7 @@ export default class IdsDataGridFormatters {
   expander(rowData: Record<string, unknown>, columnData: IdsDataGridColumn): string {
     const value: any = this.#extractValue(rowData, columnData.field);
     const button = `<ids-button tabindex="-1" class="expand-button">
-        <ids-icon slot="icon" icon="plusminus-folder-${rowData.rowExpanded === true ? 'open' : 'closed'}"></ids-icon>
+        <ids-icon icon="plusminus-folder-${rowData.rowExpanded === true ? 'open' : 'closed'}"></ids-icon>
       </ids-button>`;
     return `<span class="ids-data-grid-tree-container">${button}<span class="text-ellipsis">${value}</span></span>`;
   }